Strip flask docs from config parser
The whole system relies on `pip` to do the work. All dependencies and
operations are resolved through subprocess.

Packages are requested to be installed and qip will fetch the
dependencies and then install those one by one into the index, and
finally install the final package.

If a package is missing from the index the user will be prompted to
download it. If the user says no, the install continues with the next
dependency.