Change how target and password prompt work
Move the password and target prompt out of main context. Keep the `-t`
option in the main context to skip the prompt.

Now the --help option for commands no longer needs the target and
password

The target selection is now selection from a list. I feel this is a more
intuitive approach, as you can see the available targets.

FIX #17