Update the method to retrieve custom wiz definition
"wiz.json" will now be fetched from the python package source instead of
trying to get it from a "share" folder as we cannot set a reliable
process to make sure this bloody folder will always be there.

To do so, we need to get the module name from the top-level.txt
metadata.